requirements.txt
Flask
google-cloud-firestore
gunicorn
requests